% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_functions.R
\name{bar_plot_time}
\alias{bar_plot_time}
\title{Generate a Bar Plot for Time-Based Data Aggregations}
\usage{
bar_plot_time(
  df,
  time_var,
  price_var,
  commodity,
  pricetype,
  interactive = TRUE,
  mytitle = "Mean Price per Month for",
  x_lab = "Month",
  y_lab = "Mean Price",
  convert_axis = TRUE
)
}
\arguments{
\item{df}{Data frame containing the data to plot.}

\item{time_var}{Name of the column in df that contains the time variable (e.g., "month" or "quarter").}

\item{price_var}{Name of the column in df that contains the price variable.}

\item{commodity}{Name of the commodity (for title generation).}

\item{pricetype}{Type of the price (for title generation).}

\item{interactive}{Logical, whether to convert the plot to an interactive plotly plot.}

\item{mytitle}{Title for the plot.}

\item{x_lab}{Label for the x-axis.}

\item{y_lab}{Label for the y-axis.}

\item{convert_axis}{Logical, whether to convert the x-axis labels to a more readable format.}
}
\value{
A ggplot or plotly object, depending on the 'interactive' flag.
}
\description{
This function creates a bar plot for given dataframe, aggregated by time periods
such as months or quarters, using ggplot2 and optionally converts it to a
plotly object for interactivity.
}
