% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_functions.R
\name{calculate_label_formatting}
\alias{calculate_label_formatting}
\title{Calculate Label Formatting Based on Label Properties}
\usage{
calculate_label_formatting(df, labels)
}
\arguments{
\item{df}{Data frame containing the data to plot.}

\item{labels}{Vector of labels to be placed on the x-axis of a plot.}
}
\value{
A list containing the angle at which to display the labels, the width factor
for bar elements in bar plots, and the maximum character limit before wrapping text.
}
\description{
This function determines the optimal text formatting for plot labels based on the number
of labels and the maximum character length of the labels. It is designed to optimize
the readability of labels on plots, particularly bar plots where axis labels can
become crowded or overlap.
}
