% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_functions.R
\name{generate_welcome_message}
\alias{generate_welcome_message}
\title{Generate Welcome Message}
\usage{
generate_welcome_message()
}
\value{
An HTML formatted welcome message to be displayed in the dashboard.
}
\description{
This function generates an HTML welcome message for the Kenya Food Prices Dashboard.
}
\examples{
generate_welcome_message()
}
