% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_kenya_counties.R
\docType{data}
\name{kenya_counties}
\alias{kenya_counties}
\title{kenya_counties}
\format{
A data frame with 47 rows and 8 variables:
\describe{
  \item{ county }{  character }
  \item{ population }{  factor }
  \item{ area }{  numeric }
  \item{ pd }{  numeric }
  \item{ pdr }{  numeric }
  \item{ factor }{  numeric }
  \item{ cri }{  numeric }
  \item{ geometry }{  sfc_MULTIPOLYGON,sfc }
}
}
\source{
Source
}
\usage{
kenya_counties
}
\description{
Description.
}
\keyword{datasets}
