% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_kenya_counties_ids.R
\docType{data}
\name{kenya_counties_ids}
\alias{kenya_counties_ids}
\title{kenya_counties_ids}
\format{
A data frame with 62 rows and 2 variables:
\describe{
  \item{ county }{  factor }
  \item{ unique_id }{  integer }
}
}
\source{
Source
}
\usage{
kenya_counties_ids
}
\description{
Description.
}
\keyword{datasets}
