% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_functions.R
\name{mycolors}
\alias{mycolors}
\title{Sample Colors from the Viridis Palette}
\usage{
mycolors(n = 12)
}
\arguments{
\item{n}{The number of colors to sample from the palette.}
}
\value{
A character vector of hexadecimal color values.
}
\description{
This function randomly samples 'n' colors from the Viridis color palette,
which is designed to be perceptually uniform (in terms of human vision),
both in regular and colorblind vision. The palette generated includes 12 distinct colors.
}
\examples{
mycolors()  # Sample and print a single color from the Viridis palette
mycolors(5) # Sample and print five colors from the Viridis palette
}
