% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ke_food_prices.R
\docType{data}
\name{ke_food_prices}
\alias{ke_food_prices}
\title{ke_food_prices}
\format{
A data frame with 10647 rows and 23 variables:
\describe{
  \item{ unique_id }{  integer }
  \item{ date }{  Date }
  \item{ admin1 }{  character }
  \item{ admin2 }{  character }
  \item{ market }{  character }
  \item{ latitude }{  numeric }
  \item{ longitude }{  numeric }
  \item{ category }{  character }
  \item{ commodity }{  character }
  \item{ unit }{  character }
  \item{ priceflag }{  character }
  \item{ pricetype }{  character }
  \item{ currency }{  character }
  \item{ price }{  numeric }
  \item{ usdprice }{  numeric }
  \item{ year_month }{  character }
  \item{ year_month_date }{  Date }
  \item{ year }{  character }
  \item{ month }{  factor }
  \item{ quarter }{  integer }
  \item{ year_quarter }{  character }
  \item{ year_quarter_date }{  Date }
  \item{ county }{  character }
}
}
\source{
Source
}
\usage{
ke_food_prices
}
\description{
Description.
}
\keyword{datasets}
