% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_functions.R
\name{line_plot_prices}
\alias{line_plot_prices}
\title{Line Plot Prices}
\usage{
line_plot_prices(mean_prices, commodity, pricetype)
}
\arguments{
\item{commodity}{The commodity to filter the data by.}

\item{pricetype}{The price type to filter the data by.}

\item{data}{A data.table containing the food price data.}

\item{palette}{The color palette to use for the plot. Default is "Set1".}
}
\value{
A plotly object containing the line plot.
}
\description{
This function creates a line plot of prices for a specified commodity and price type.
}
